Add libgfortran versioning into Platforms.
This adds a new `libgfortran_abi` portion to all platforms.  When
parsing and generating triplets, a new tag is added onto the end (if it
is not present, it is assumed there is no fortran code present within
the tarball, and it therefore has no dependence on a gfortran ABI).

Tarball installation is now not quite as simple as finding the
appropriate key within a dictionary; it is now (theoretically) possible
that multiple keys within a `download_info` dict can apply to the same
julia installation.  (E.g. if the julia installation has no gfortran
already linked into it, then a libgfortran3 linked tarball and a
libgfortran4 linked tarball are both applicable).   To resolve
ambiguities, the `choose_download()` function is now available and
should be used to pull out the `url` and `hash` from a `download_info`
dictionary.