Add copyderef workaround for symlink-unfriendly filesystems
We allow two methods to enable this workaround: Setting
`BINARYPROVIDER_COPYDEREF=true` in your environment to force it on, and
also auto-detection that runs every time you try to `unpack()` a
tarball.  If auto-detection triggers (e.g. we attempt to create a symlink
and it fails, and symlinks are able to be created within `tempdir()`)
we automatically enable `copyderef` mode, which will extract tarballs into
`tempdir()`, then copy the files from that temporary directory into the
true destination, collapsing symlinks and skipping broken links.